Extract ArrayInt64 utilities
